pyproject.toml
[project]
name = "agent-second-brain"
version = "1.0.0"
description = "Voice-first personal assistant for Telegram with Obsidian and Todoist integration"
readme = "README.md"
authors = [
    { name = "Agent Second Brain Contributors" }
]
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.0",
    "deepgram-sdk",
    "fastapi>=0.135.1",
    "httpx",
    "pydantic>=2.0",
    "pydantic-settings",
    "python-multipart>=0.0.22",
    "todoist-api-python>=3.1.0",
    "uvicorn[standard]>=0.41.0",
    "yt-dlp>=2026.3.3",
    "pymupdf>=1.24.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "trafilatura>=2.0.0",
    "anthropic>=0.50",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "d_brain"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
